写作宝缓存

+++ src/main/java/com/stonedt/intelligence/vo/BindParamsVo.java
@@ -13,7 +13,7 @@ public class BindParamsVo {
     /**
      * 用户id
      */
-    private long userId;
+    private int userId;
 
     /**
      * 认证id
